backend-app/requirements.txt
fastapi==0.109.2
numpy==1.26.4
pandas==2.2.1
protobuf==4.25.3
python-dotenv==1.0.1
